backend/backend-service-for-image-analysis-kiln/requirements.txt
google-genai
google-cloud-storage
Flask
Flask-Cors